migrate all api to app routes
